+++ source/components/table/table.js
@@ -19,6 +19,8 @@ const TABLE_COL_OPTIONS = {
 };
 
 ((Metro, $) => {
+    "use strict";
+
     let TableDefaultConfig = {
         caption: "",
         // cardMode: "",