+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 5.1.18
 
 + [x] Fixed `bg-*-clown` family colors
-+ [x] Corrected saturated and darken colors 
++ [x] Corrected saturated and darken colors
++ [x] Added animation to the linked block connector